refactor: split AnalysisPage
prepare for single-label and multi-label split in labels
