+++ src/pages/analysis/SingleLabelAnalysis.tsx
@@ -1,0 +1,14 @@
+export function SingleLabelAnalysis() {
+  return (
+    <main className="max-w-[920px] mx-auto px-14 pt-18 pb-32 font-serif text-on-canvas">
+      <header className="pb-3.5 border-b border-edge-warm flex items-baseline justify-between text-stone text-[12px]">
+        <span style={{ letterSpacing: '0.08em' }}>CHATSIGHT · FINDINGS · single-label runs</span>
+        <span style={{ fontStyle: 'italic' }}>placeholder · Phase 0</span>
+      </header>
+      <h1 className="text-5xl font-medium tracking-tight mt-9 mb-2">Findings</h1>
+      <p className="italic text-[17px] text-tertiary max-w-[560px]">
+        (SingleLabelAnalysis placeholder &mdash; filled in during Phase 2.)
+      </p>
+    </main>
+  )
+}
